renamed global_straggler_detector
